Split App.vue into surfaces, a sidebar, and four composables
App.vue was 621 lines: the window's whole navigation model, every keyboard
command, two polling loops, the zoom subscription, and both work surfaces
written out in full. It is now 179 and reads as what it is — the shell that
composes the rest.

- useWorkbench: which surface is showing and everything that changes it,
  including the target subscriptions and what the window opens with
- useReviewKeyboard: every binding, the cursor it moves, and the tree jump
- useBackgroundRefresh: the review poll, the service poll, and refresh on focus
- useWindowZoom: the level, its subscription, and the clamp
- LocalSurface / ReviewSurface: the two work surfaces, with the review
  forwarding the diff's two keyboard commands
- WorkbenchSidebar: the tree, its splitter, and the scrollbar's fade
- styles/work-surface.css: what both surfaces share, on the tree.css pattern
